requirements/base.txt
Django
djangorestframework
django-filter
django-extensions
psycopg2-binary
